refactor: restructure docs-extractor for modular architecture
- Removed the ComponentAnalyzer and ExternalTypeResolver classes, replacing them with standalone functions in new files.
- Introduced PropsAnalyzer for extracting props information from TypeScript types.
- Created a new type-extractor module to handle TypeScript program creation and component type extraction.
- Updated index.ts to reflect new module structure and exports.
- Removed legacy VanillaExtractAnalyzer and replaced it with utility functions for CSS file handling.
- Cleaned up unused utility functions and consolidated prop extraction logic.
- Updated tsconfig.json for improved path resolution.
