+++ packages/docs-extractor/src/bin/cli.ts
@@ -3,9 +3,9 @@ import * as path from 'path';
 import yargs from 'yargs';
 import { hideBin } from 'yargs/helpers';
 
-import { extractComponentTypesFromFile } from '~/type';
-import { RunOptions } from '~/types/types';
-import { createComponentData, ensureOutputDirectory, writeComponentDataToFile } from '~/utils';
+import { extractComponentTypesFromFile } from '../type-extractor';
+import { RunOptions } from '../types/types';
+import { createComponentData, ensureOutputDirectory, writeComponentDataToFile } from '../utils';
 
 export function createCliCommand(runFunction: (options: RunOptions) => Promise<void>) {
     return yargs(hideBin(process.argv))
